requirements.txt
flask
python-dotenv
pillow
beautifulsoup4
faiss-cpu
sentence-transformers
google-generativeai
numpy
